+++ src/core/ddsc/src/dds__rhc.h
@@ -12,60 +12,81 @@
 #ifndef _DDS_RHC_H_
 #define _DDS_RHC_H_
 
+#include "dds/ddsrt/static_assert.h"
+#include "dds/ddsi/q_rhc.h"
+#include "dds__types.h" /* for dds_readcond */
+
 #define NO_STATE_MASK_SET   (DDS_ANY_STATE + 1)
 
 #if defined (__cplusplus)
 extern "C" {
 #endif
 
-struct rhc;
-struct dds_qos;
-struct ddsi_serdata;
-struct ddsi_tkmap_instance;
-struct proxy_writer_info;
+struct dds_rhc;
 
-DDS_EXPORT struct rhc *dds_rhc_new (dds_reader *reader, const struct ddsi_sertopic *topic);
-DDS_EXPORT void dds_rhc_free (struct rhc *rhc);
+typedef int (*dds_rhc_read_t) (struct dds_rhc *rhc, bool lock, void **values, dds_sample_info_t *info_seq, uint32_t max_samples, uint32_t mask, dds_instance_handle_t handle, struct dds_readcond *cond);
+typedef int (*dds_rhc_take_t) (struct dds_rhc *rhc, bool lock, void **values, dds_sample_info_t *info_seq, uint32_t max_samples, uint32_t mask, dds_instance_handle_t handle, struct dds_readcond *cond);
+typedef int (*dds_rhc_takecdr_t) (struct dds_rhc *rhc, bool lock, struct ddsi_serdata **values, dds_sample_info_t *info_seq, uint32_t max_samples, uint32_t sample_states, uint32_t view_states, uint32_t instance_states, dds_instance_handle_t handle);
 
-DDS_EXPORT uint32_t dds_rhc_lock_samples (struct rhc *rhc);
+typedef bool (*dds_rhc_add_readcondition_t) (struct dds_readcond *cond);
+typedef void (*dds_rhc_remove_readcondition_t) (struct dds_readcond *cond);
 
-DDS_EXPORT bool dds_rhc_store  (struct rhc * __restrict rhc, const struct proxy_writer_info * __restrict pwr_info, struct ddsi_serdata * __restrict sample, struct ddsi_tkmap_instance * __restrict tk);
-DDS_EXPORT void dds_rhc_unregister_wr (struct rhc * __restrict rhc, const struct proxy_writer_info * __restrict pwr_info);
-DDS_EXPORT void dds_rhc_relinquish_ownership (struct rhc * __restrict rhc, const uint64_t wr_iid);
+typedef uint32_t (*dds_rhc_lock_samples_t) (struct dds_rhc *rhc);
 
-DDS_EXPORT int
-dds_rhc_read(
-        struct rhc *rhc,
-        bool lock,
-        void ** values,
-        dds_sample_info_t *info_seq,
-        uint32_t max_samples,
-        uint32_t mask,
-        dds_instance_handle_t handle,
-        dds_readcond *cond);
-DDS_EXPORT int
-dds_rhc_take(
-        struct rhc *rhc,
-        bool lock,
-        void ** values,
-        dds_sample_info_t *info_seq,
-        uint32_t max_samples,
-        uint32_t mask,
-        dds_instance_handle_t handle,
-        dds_readcond *cond);
+struct dds_rhc_ops {
+  /* A copy of DDSI rhc ops comes first so we can use either interface without
+     additional indirections */
+  struct rhc_ops rhc_ops;
+  dds_rhc_read_t read;
+  dds_rhc_take_t take;
+  dds_rhc_takecdr_t takecdr;
+  dds_rhc_add_readcondition_t add_readcondition;
+  dds_rhc_remove_readcondition_t remove_readcondition;
+  dds_rhc_lock_samples_t lock_samples;
+};
 
-DDS_EXPORT void dds_rhc_set_qos (struct rhc * rhc, const struct dds_qos * qos);
+struct dds_rhc {
+  union {
+    const struct dds_rhc_ops *ops;
+    struct rhc rhc;
+  } common;
+};
 
-DDS_EXPORT bool dds_rhc_add_readcondition (dds_readcond * cond);
-DDS_EXPORT void dds_rhc_remove_readcondition (dds_readcond * cond);
+DDSRT_STATIC_ASSERT (offsetof (struct dds_rhc, common.ops) == offsetof (struct rhc, ops));
 
-DDS_EXPORT int dds_rhc_takecdr
-(
-  struct rhc *rhc, bool lock, struct ddsi_serdata **values, dds_sample_info_t *info_seq,
-  uint32_t max_samples, uint32_t sample_states,
-  uint32_t view_states, uint32_t instance_states,
-  dds_instance_handle_t handle
-);
+DDS_EXPORT inline bool dds_rhc_store (struct dds_rhc * __restrict rhc, const struct proxy_writer_info * __restrict pwr_info, struct ddsi_serdata * __restrict sample, struct ddsi_tkmap_instance * __restrict tk) {
+  return rhc->common.ops->rhc_ops.store (&rhc->common.rhc, pwr_info, sample, tk);
+}
+DDS_EXPORT inline void dds_rhc_unregister_wr (struct dds_rhc * __restrict rhc, const struct proxy_writer_info * __restrict pwr_info) {
+  rhc->common.ops->rhc_ops.unregister_wr (&rhc->common.rhc, pwr_info);
+}
+DDS_EXPORT inline void dds_rhc_relinquish_ownership (struct dds_rhc * __restrict rhc, const uint64_t wr_iid) {
+  rhc->common.ops->rhc_ops.relinquish_ownership (&rhc->common.rhc, wr_iid);
+}
+DDS_EXPORT inline void dds_rhc_set_qos (struct dds_rhc *rhc, const struct dds_qos *qos) {
+  rhc->common.ops->rhc_ops.set_qos (&rhc->common.rhc, qos);
+}
+DDS_EXPORT inline void dds_rhc_free (struct dds_rhc *rhc) {
+  rhc->common.ops->rhc_ops.free (&rhc->common.rhc);
+}
+DDS_EXPORT inline int dds_rhc_read (struct dds_rhc *rhc, bool lock, void **values, dds_sample_info_t *info_seq, uint32_t max_samples, uint32_t mask, dds_instance_handle_t handle, struct dds_readcond *cond) {
+  return rhc->common.ops->read (rhc, lock, values, info_seq, max_samples, mask, handle, cond);
+}
+DDS_EXPORT inline int dds_rhc_take (struct dds_rhc *rhc, bool lock, void **values, dds_sample_info_t *info_seq, uint32_t max_samples, uint32_t mask, dds_instance_handle_t handle, struct dds_readcond *cond) {
+  return rhc->common.ops->take (rhc, lock, values, info_seq, max_samples, mask, handle, cond);
+}
+DDS_EXPORT inline int dds_rhc_takecdr (struct dds_rhc *rhc, bool lock, struct ddsi_serdata **values, dds_sample_info_t *info_seq, uint32_t max_samples, uint32_t sample_states, uint32_t view_states, uint32_t instance_states, dds_instance_handle_t handle) {
+  return rhc->common.ops->takecdr (rhc, lock, values, info_seq, max_samples, sample_states, view_states, instance_states, handle);
+}
+DDS_EXPORT inline bool dds_rhc_add_readcondition (struct dds_readcond *cond) {
+  return cond->m_rhc->common.ops->add_readcondition (cond);
+}
+DDS_EXPORT inline void dds_rhc_remove_readcondition (struct dds_readcond *cond) {
+  cond->m_rhc->common.ops->remove_readcondition (cond);
+}
+DDS_EXPORT inline uint32_t dds_rhc_lock_samples (struct dds_rhc *rhc) {
+  return rhc->common.ops->lock_samples (rhc);
+}
 
 #if defined (__cplusplus)
 }
