+++ src/core/ddsc/src/dds_init.c
@@ -272,12 +272,6 @@ void ddsi_plugin_init (void)
   ddsi_plugin.builtintopic_is_visible = dds__builtin_is_visible;
   ddsi_plugin.builtintopic_get_tkmap_entry = dds__builtin_get_tkmap_entry;
   ddsi_plugin.builtintopic_write = dds__builtin_write;
-
-  ddsi_plugin.rhc_plugin.rhc_free_fn = dds_rhc_free;
-  ddsi_plugin.rhc_plugin.rhc_store_fn = dds_rhc_store;
-  ddsi_plugin.rhc_plugin.rhc_unregister_wr_fn = dds_rhc_unregister_wr;
-  ddsi_plugin.rhc_plugin.rhc_relinquish_ownership_fn = dds_rhc_relinquish_ownership;
-  ddsi_plugin.rhc_plugin.rhc_set_qos_fn = dds_rhc_set_qos;
 }
 
 //provides explicit default domain id.
